Porting baseline test for vgen.
This is a rewrite of the original infrastructure so it works with the
different testdata model which need to be used in git compared to
google3.

Note: at this point protoc needs to be in the path, as the tests need
to compile protos. This need to be fixed later.

Change-Id: If17d7dc62ece86e12906a2e373a7de2395ce9101