Correctly handle protobuf's maps and bytes tokens
Protobuf's map<A,B> type gets converted implicitly to MapEntry
with repeated cardinality.
When protobuf generates Java code however,
it changes the type back to Map<A,B>.
We were generating code using List<MapEntry>.
This fails, since a list of entries does not conver to a map.
This patch makes us correctly use Map in this situation.

We were incorrectly casting all page streaming tokens to String.
This patch correctly cast to ByteString when appropriate.

We were calling "new ByteString()",
which is incorrect since ByteString is an interface.
This patch correctly calls ByteString.EMPTY.

Change-Id: I442eec6f7547ba4ce77f505221b9b34b19ff011f