Generate a factory class for simple client creation
The ChannelFactory class used here does not currently exist in source control, and its location
is also TBD, but it will exist before the generated code is ever checked in. The host/port/scopes
may also be extracted to something akin to the Java ServiceApiSettings class, but even in that
case the per-service factory class would still exist to expose the default settings and convenience
initializers.

Change-Id: I5d8321f5f7e3e5723ad98f95142c3daad376b061