Fix empty PathTemplate bug
Generate no PathTemplate class if there are no path templates.

Previously, if no PathTemplates were specified in the config, VGen
generated a PathTemplate class with an empty body, which is a Python
syntax error.

Also, add a baseline test for this case.

Additionally, add some checks to avoid generating whitespace when
certain optional sections are not present. As a general rule, I
suggest always wrapping optional sections in the snippets with {@if}
to avoid generating extraneous whitespace.

Change-Id: I23cbcb52e66c69b6ed6170519320ad2f1c9855b4