Python VGen update for bundling
Introduces bundling into Python VGen. To accomodate bundling, the
defaults/options system has been reworked; ApiCallDefaults is removed.
Instead, a dictionary is maintained which associates defaults to each
method; the values in the dictionary are then merged with any
CallOptions passed in while a call is being made.

A corresponding change to GAX is here:
  https://github.com/googleapis/gax-python/pull/28/

Note that it is not currently possible to override the bundling
defaults when calling the API constructor; a later change will add
the ability to pass in a dictionary mapping method names to bundlers
in order to provide this functionality.

Change-Id: I434a6026882fa57ec0d4672bff8be4f4102c4a82