Remove context manager support for API objects
Context manager is no longer required due to gRPC updates.

In particular, the gRPC stub should now be garbage collected correctly
when it goes out of scope, and clean up the relevant resources. This
means we don't have to explicitly free the stub through the __exit__()
function.

Pre-push hook installed.

Change-Id: Ic6068d231a7d35e5e8a5e19620c0b33757897532