Stop test from spamming /tmp
TestDataLocator::getTestDataAsFile copies data into
a new file in /tmp then returns the Path to the file,
presumably to prevent accidental overwrite of the files in the repo.

In case of testing discovery code fragment generation, we only want
to read the file. This change makes the test read the file directly,
thus avoiding the file copy into /tmp

Pre-push hook installed.

Change-Id: I98ecc0a189952f388132e2bb960142a3bc3c880f