Refactor language providers
Each language provider instance now represents a single template
(i.e. code, doc, or fragment) for a given language. It has an internal
strategy object (*View) to encapsulate template-specific generation
details. The structure of the language provider is simplified to have
a single generate method, corresponding to the generation of a single
template.

The structure of the GAPIC config YAML files has changed to push the
snippet definitions down to a per-template level, to avoid the
explosion of keys for various types of snippet files. A template
represents a self-contained, independently-runnable phase of code
generation, consisting of a language provider, a view, and a set of
templates.

Pre-push hook installed.

Change-Id: Id682441e1b8a3f3209974e0d3ee3642d5a6ce610