Fix Go Discovery Generation
The bug was introduced with Python Discovery generation, and I missed
it.

This commit reverts the change to DiscoveryImporter: Now it will only
report the return type of "empty$" if the return type is completely
missing from the Discovery doc.

Most languages can ignore return type "Empty", so the string "Empty" is
especially handled. Go cannot ignore "Empty" type and is special-cased.