Update Go paging iterators (#220)
Update Go paging iterators

The slight differences between Next and NextPage were hard to get right
using the existing division of labor. Refactored so that apiCall does
not deal with when to return Done, it just calls the List method and
sets some iterator fields. Deciding when to return Done and what to
return along with it is left to Next and NextPage.

This change was tested against the iterator tests of the logging client.