Get rid of '..' path in templates
The path '..' is special to UNIX.
When running under non-UNIX-like file systems,

  @extends "../common.snip"

does not work since the directory ".." loses its special meaning.

"Non-UNIX-like file systems" may also appear on UNIX systems.
If the project is packaged into a JAR for distribution,
the '..' directory will still fail, since a JAR is a Zip archive,
and a Zip archive does not have '..' directories.

The '..' was used in the first place to reuse template code stored in
"codegen/common.snip", while the resource root is in subdirectories,
like "codegen/go".

This commit changes the resource root of all template renderings to
"codegen".
In this way, "codegen/common.snip" can simply be used by

  @extends "common.snip"

and files in subdirectories can be used by

  @extends "subdir/foo.snip".

Fixes #260.