Refactoring PHP generation to use the MVVM pattern
This was prototyped in https://github.com/googleapis/toolkit/pull/268
for both Java and PHP; the present PR only includes PHP and shared
stuff.

The MVVM pattern uses the following steps:

1. The Model is converted to a ViewModel, which is a set of classes
containing the data to be used when rendering the templates.
2. The snippet engine renders the output using only data from the
ViewModel structure.

Other changes with the MVVM model:

* There is no context object when generating under the MVVM model. All
  data must be prepared beforehand.
* No casing changes are done in templates; all identifiers must be
  composed and cased correctly inside the view model.

There are a number of new utility classes created:

* Name: represents an identifier name which is casing-aware.
* TypeName: Represents a simple or complex type and keeps track of the
  aliases for the contributing types.
* TypeTable: manages the imports for a set of fully-qualified type
  names.
* SurfaceNamer: provides language-specific names or other strings.

There is also a change unrelated to MVVM:

* Using TreeSet for auth scopes to give a consistent ordering