DO NOT COMMIT: Prototype getRepresentativeMethod
Issue #245 is caused by the way we display a method sample in API doc.
We simply look for the first flattened method,
assume (most likely without meaning to) that it's not page-streaming,
and create a sample for it.

The issue itself can be resolved by properly detecting whether or not
the method page-streams. This can be done after MVVM to avoid duplicate
work.

But this begs the question: what the heck do we do if there is no
flattened method?

This commit prototypes a way to find a suitable method with the
following observations.
* we prefer flattened method; that's what users will probably use
* we prefer non-page-streaming methods; their samples are shorter
* we prefer methods with fewer arguments; initializations are shorter