Ignore streaming methods.
Currently Gapic generator does not support gRPC-streaming methods,
we will skip streaming methods for the time being (otherwise
the generated code will behavior incorrectly or lead to compile
errors).

Note that this skip can't be done solely on the configuration
yaml (i.e. not-listing methods in the yaml file), because
many of the generators silently assume the existence of
MethodConfig instances and end up with NullPointerExceptions.

Rather than supporting missing MethodConfig cases, I believe this
special-handling of gRPC-streaming would be better because
anyways this type of special-handling will be necessary when we start
supporting gRPC-streaming methods for a specific language without
affecting other languages.