Rename Go Imports and Properly Wrap API Doc
This commit further simplifies Go import names.
Prior to this, we take the import path such as "googleapis/library/v1"
and replace the slashes with underscore.

This commit renames the primary proto package to just "pb".
Other proto packages are named in the format of "foopb".

Top-level package documents were not line-wrapped.
This commit also line wraps them.