Incremental dependency bump
This is only an incremental bump to limit the scope of changes.

* protobuf-gradle-plugin: 0.7.2-SNAPSHOT to 0.7.3 (getting off of a
  snapshot dependency)
* protobuf/protoc: 3.0.0-beta-2 to 3.0.0-beta-3 (consistent with the
  version we are using in the generated code stack)

Also, renaming "private" snapshot repo references to "toolkit". This
means the username and password entries in each dev's gradle.properties
file need to be updated, if they want to use 'uploadArchives'. The
basic problem is that multiple repos might use the same
keys (e.g. api-compiler), and thus would clash with each other.

* privateOssrhUsername to privateToolkitUsername
* privateOssrhPassword to privateToolkitPassword