Go: Make sure clientNamePrefix doesn't break on new Interfaces
Go has a logic of dropping service names when they are not necessary.
Eg, instead of "library.NewLibraryClient",
we just say "library.NewClient".

Previously we drop the service name in two situations
- When the service is the only one defined in the package
- When the name of the service is the same as the name of the package

This commit removes the first condition,
as it might break client libraries when new services are added
to the package later on.