Go: Properly handle new go_package format
I'm currently working to add go_package option in appropriate places.
The format of the option is in form
  "path/to/pkg;localname"
which is different than the one we are currently using.

This commit lets us handle both.
The logic handling the old version can be deleted after the go_package
change lands in googleapis.