Fix the pattern of param name conflicting with language keywords.
Fixes #618.

As is seen in the bug, a name pattern for cloud functions
contain 'function' name which becomes to be a parameter of
a function and therefore causes syntax error on JS.

Potentially the same problem could happen in many other languages.
This patch includes fixes on those potential problems.