Fix the pattern of enums.
The enum values should be tracked as packageName.Path.To.Enum.VALUE
in NodeJS.

In order to refer to packageName in the architecture, I found
that the current package_name structure for nodejs gapic config
does not fit well into the current MVVM architecture.

So I changed the expected format for the 'package_name' field
in the gapic config as 'apiname.version'. This makes the name
being not a package name of NPM, but probably okay for now,
and it also indicates how the module will be accessed in the
code.

For simplicity, I put the scope part `@google-cloud/` into
`domain_layer_location`.

Fixes #705.