Port GAPIC metadata generation for Python
Move the functionality of generating Python GAPIC package
metadata from Packman to GAPIC. To do so, introduce a new package
metadata configuration format, which lives at the same level as the
GAPIC config in terms of input/test input. This new config file
fixes #848.

Sorry, this PR is big, but a lot of the lines changed are test data.

The following follow-up PRs are planned:
- A PR to Artman, introducing a task to generate the
  new package metadata config from the artman config, and removing the
  Python Packman task.
- A PR to Toolkit, finishing the port of Packman templates with the
  "docs" templates, which automate documentation generation.
- A PR to Toolkit, replacing all inputs to GrpcMetadataGen with the
  new package metadata config, which addresses #780
- (Potentially) A PR to Toolkit investigating the consolidation of
  gRPC and GAPIC Python package metadata snippet templates. Note that
  Packman currently uses separate templates for gRPC and GAPIC.