make Go tests forward-compatible (#862)
When mocking servers, we generate all methods to satisfy the server
interface. On the other hand, adding more methods to services is
considered non-breaking change in protobuf.
This breaks the existing tests though, since the newly added method
won't be implemented.

This change fixes the problem by embedding a nil server interface
in the mock. This way the mock always implements the interface,
even though some methods might not actually be implemented.
The methods on both the client and the mock will just be added
the next time the generator is run.