Plug Python into MVVM machinery
This is prior to fixing the snippets to not depend on elements of the
Model; instead getting what they need from ViewModel objects. It’s not
really MVVM yet, but it allows tests to pass while migrating, so it can
be done incrementally (and thus is easier to reason about and review).