Port Python Sphinx doc files from Packman (#924)
This is a translation of the mustache files here:

  https://github.com/googleapis/packman/tree/master/templates/gax/python/docs

to toolkit templates.

Note that these templates require knowledge of the Python modules produced
during earlier stages of GAPIC generation. To do so, the PR passes the
`GapicProvider`s for these stages directly to the PackageMetadataTransformer,
and calls `generate()` on those providers to determine their output. After
Python is on MVVM, we can avoid actually running `generate()` twice by passing
the relevant transformers, and get the output module filenames from the
ViewModels they produce without actually rendering the templates twice.