update Go header
A couple things of note:

The GRPC library doesn't expose a version yet, so that's not available
right now

The version string is in YYYYMMDD format. Rationale:
The semver spec states that "Once a versioned package has been released,
the contents of that version MUST NOT be modified". Since "go get" just
downloads the master branch, we are technically modifying the version
every time we submit code. If our users use the version package for
their own telemetry, they might be unpleasantly surprised.