fix Go top-level enum
Go names enums in two different ways.
If the enum is nested in a message, the format is <messagetype>_<enum value>,
respecting the C++ scoping rule used by protobuf,
where enum values are scoped at the same level as the
enum type, not as its children.
On the other hand, if the enum is at top-level,
there is no parent message,
and the format is <enum type>_<enum value>