go: small codegen fixes
Previously we unit-test error cases with error Canceled.
However, since pubsub is configured to retry on this error,
we retry forever and never gives up.
This commit changes the error to PermissionDenied instead,
since it seems unlikely that there's a retry case for it.

This commit also addresses a naming inconsistency.
Unlike Java, Go spells "canceled", "canceling", etc
with only one L.