Use placeholders for strings that are URL vars
Parameters which are URL variables have the attribute `location` set to `path`
in the Discovery file. This commit adds support for that field, and adds a
non-empty placeholder format for URL variable strings.

The reasoning is that samples executed with URL variables that are empty
strings will return a 404 error, because client libraries combine contiguous
slashes.

For example, the following sample for the method `/foo/{bar}/baz`:

    bar = "" service.get_baz(bar)

will send a request to `/foo/baz` instead.

This commit will generate the following sample instead:

    bar = "my-bar" service.get_baz(bar)

which correctly sends the request to `/bar/my-bar/baz`.

Note that URL variables are always strings, but parameters which have a
non-string `format` specified are not affected by this change.