Use trailing commas (#1171)
This commit adds a trailing comma to the last item of each object. Due to our
limitations in snippet language, and the fact that this code is autogenerated,
it's easier to use this style than try to ensure that the last item in an
object has no comma following it.

This commit also achieves consistency, as the samples are currently
inconsistent because all optional request body fields are generated with
trailing commas.