go: fix name prefix
In Go, if a service has the same name as a package,
we omit the service name.
For example, in the "foo" package, the client for service "Foo"
is "foo.Client" not "foo.FooClient".

Previously, this did not work for multi-word package name:
In the "foobar" package, the client for service "FooBar" was
named "foobar.FooBarClient".

This commit properly names it "foobar.Client".

Some note worthy points:
- This could theoretically create name conflicts. Both "FooBar"
  and "FoobAr" will both be named "foobar.Client". However,
  I don't think this will be a problem in practice.
- This will rename videointelligence.VideoIntelligenceClient to
  videointelligence.Client. Since the API is only in alpha,
  I think this change in acceptable.