all: make interface renaming work for go
Go uses the interface name as parts of many other names
in the client.
This commit makes the rename available to SurfaceNamer implementations.
The renaming itself is not "threaded through" yet.
Since this change is already large, I plan to do this in a follow-up commit.

An alternative to this is to give GoSurfaceNamer the rename map.
This would put renaming logic in two places, but it will isolate the change
to Go and leave other languages alone.
Reviewers, please let me know if you'd rather I do that instead.

I tested that Go and Java googleapis clients remain the same after this change.