Add overview.rb file to Ruby docs (#1588)
This overview serves as the landing page for documentation for
GAPIC-only APIs. Previously, it was not rendered due to the same
feature shortcoming in the jsondoc tool that motivated

  https://github.com/googleapis/toolkit/pull/1544

Retain readme text in the version index to serve as a per-version
code level example, but remove installation instructions, since
installation is not API version-dependent.