Add PHP smoke test (#1610)
This commit adds a smoke test snippet and transformer for PHP.

As part of that addition, the generated test directory for the PHP GAPIC
client is reorganized into `tests/unit`, which contains unit tests, and
`tests/system`, which contains the smoke test. This also mirrors the
layout on google-cloud-php.

PhpGapicSurfaceTestTransformer.java is mostly rewritten to improve code
clarity.