Delete Discovery samplegen (#1640)
The contents of this repo as of 2017-10-20 12:40 are captured in
github.com/googleapis/discovery-artifact-manager/toolkit/. This is a
one-time sync to cut development costs on the toolkit repo, which will
continue to evolve. We expect few or no changes to the Discovery
samplegen code, which will henceforth only live in the
discovery-artifact-manager repo, not in toolkit.