Allow user initialization TODOs in samples and apply this to Ruby (#1813)
In some autogenerated samples, additional user input is required to
obtain working code if the sample has not been fully configured. In
the case where a required field is not configured, add a comment to
the sample code.

This PR adds the comment to the view model as well as the Ruby templates.