all: fix oneof clash in generated test
We previously initialize all primitive fields in tests.
When we see multiple primitive fields in a oneof group,
we init all of them.

This commit dedupes by oneof group, so we only init only one.

Also, add "context" to Go list of "keywords". It's so commonly used
it might as well be reserved.