Generate empty standalone sample files in Java
(This PR is still in progress but I want to get early feedback as I
continue working on it.)

With these changes, stubs of Java standalone samples are generated for
methods whose GAPIC configs are configured to do so via the method's
"sample_value_sets" and "samples" fields.

Before this PR goes in I plan to do the following:
- add equivalent support for Python (I want to make sure my approach works for both dynamic and static languages)
- fill in a few of the stubs used for Java and Python
- add more tests

In subsequent PRs, I will pass to the .snip files in a usable form the
actual sample value sets for each method sample

I touch the path generator for other languages to account for a
modified interface. Let's not worry yet about the generated location of the
samples for those. At any rate, for all languages artman can be
configured to move the samples to the rigth final directory for the user.