Handle map return type (#1900)
In Discovery doc schemas, the additionalProperties field is a dynamically-keyed map type. For now, assume the keys are Strings. There are pageable methods, e.g. compute.addresses.aggregatedList, that must page over this Map<String, ? extends ApiMessage> type.

In discogapic configgen, make page_streaming.response.resources_field the name of the field in the response containing the list of resources belonging to the page, as the config.proto has dictated.

For paged methods, scrap the concept of a list of field getter methods to get from a response object to the paged resource object. Assume the paged resource object is a direct field of the response object.

PageStreamingConfig constructor is now same regardless of model type (e.g. proto/discovery)