Generate stub standalone sample files in Java and Python (#1874)
With these changes, **stubs** of Java and Python standalone samples are generated for
methods whose GAPIC configs are configured to do so via the method's
`sample_value_sets` and `samples` fields. The relevant configuration in those samples get passed to the sample templates via the `initCode` fields of the sample view models.

Note that these changes are backwards-compatible: since currently released APIs do not have samples configured, no sample files are generated for those.

Note also that the stubs produced by this PR use the existing in-code samples, which are not standalone.

I touch the path generator for other languages to account for a
modified interface. Let's not worry yet about the generated location of the
samples for those. At any rate, for all languages artman can be
configured to move the samples to the right final directory for the user.

High-level overview:

   - Main data structures:
     - Since API methods in general contain samples (in-code samples in the library surface itself, standalone samples in the standalone sample files), `{StaticLang,OptionalArray}MethodView` both now contain a getter `samples()` which returns a `List<MethodSampleView>`
     - `MethodSampleView` contains the view model for the method-level specific sample output, including the relevant `InitCodeView` (and `ApiMethodView.initCode` will be deprecated in favor of the new sample config)
     - `StaticLangSampleClassView` and `DynamicLangSampleView` contain the top-level view model for the sample classes/files that show the usage of a particular method in the library surface.

   - High level flow
     - Configuration options get read into `SampleSpec`, now a part of `MethodConfig`
     - Transformation flow:
        `<Lang>Gapic{Surface,Samples}Transformer`: Entry point for generating the top level client or sample artifacts. Generates `StaticLang{Api,SampleClass}View`/`DynamicLang{Api,Sample}View`
         -> `<Lang>MethodViewGenerator`: Logic for determining which method type (and eventually which calling forms) to select
         -> `{Static,Dynamic}LangApiMethodTransformer`: Generating the `ApiMethodView`s, including samples
         -> `SampleTransformer`: Generating the `MethodSampleView`s appropriate to the artifact being produced

   - Rendering entry points for samples are now `<LANG>/standalone_sample.snip`