go: don't overwrite the List requests (#2019)
In list calls, the iterator might make multiple RPCs with different page token to fetch items. Previously, we use the request object the user supplied and change its PageToken field in place. This behavior is probably unexpected and certainly unintentional.

This PR fixes the problem by making a copy of the request object for the iterator's own use and never modify the one the user passes in.

This is technically a breaking change: people might be counting on the request being modified, but this seems extremely unlikely.