go: don't generate empty LRO Metadata() (#2051)
Most methods we generate have signatures like this:

  func (*Client) Method(context.Context, *pb.Request) (*pb.Response, error)

If the response is an empty type, it's not terribly useful, so we drop
it, making the signature:

  func (*Client) Method(context.Context, *pb.Request) error

Since we never write the name of the Empty type,
we have a special logic to never import the "emptypb" package
so we don't get the imported-but-not-used error.

The problem is that if a long-running operation has an Empty metadata,
we still create the

  func (*Operation) Metadata() (*emptypb.Empty, error)

method. This does not compile, because now we're using emptypb without
importing.

The solution is to simply not generating the Metadata() method at all,
since it doesn't contain any useful information anyway.