Account for java_multiple_baselines=false in naming grpc elements (#2113)
When java_multiple_files=false in the proto API definition, the generated proto files are put under a single outer file. However, the generated grpc class names are not put under a single outer file.

The gapic-generator should also match the grpc import paths by not putting the grpc classes under an outer Service file. In the baseline tests, TaggerProto is the service file in which proto classes are put, but grpc classes should not be put.