Enable naive server streaming retries.
This enables retries that were added to gax in
https://github.com/googleapis/gax-java/pull/463.
Server streaming api can be treated the same as unary apis before the
first response is received. Allowing GAPIC clients to automatically
retry RPCs. This PR exposes the retry settings that were configured in
gapic yaml file to the java stub settings.

This will affect the codegen for 3 clients:
- firestore - will enable retries for batchGetDocuments & runQuery.
  Which will bring its behavior closer to the golang client
- bigtable - the handwritten overlay will overwrite these settings
- spanner - doesn't currently use gapic client, so will ignore the new
  settings