[Proto Annotations] Enable passing in --package flag (#2248)
The --package flag is necessary when only a descriptor set is given (no service yaml, no gapic config, no package2.yaml). It specifies the proto package e.g. google.pubsub.v1 for the source protos that we actually want to make a library for, as opposed to imported packages like google.api.http.