Don't generate methods that aren't present in GAPIC config (#2381)
While we are in the migration process off GAPIC config, we shouldn't generate methods that are in the protofile but not in the GAPIC config because that would be a breaking change. 

Also don't generate retry codes for methods that aren't defined in the GAPIC config.

Add a method to the test library.proto but not to its test GAPIC config to test that it won't be present in the generated surface.