Run Artman smoke tests as part of CircleCI (#2443)
Run the full suite of artman smoke tests. This does even more tightly couple gapic-generator to artman for now, but it will save us time down the road when we go to release artman and gapic-generator has broken artman.

Soon after this, we can put in a CircleCI job filter so these smoke tests only run on the master branch, given that they take a long time (~40 minutes).