[Proto Annotations] Paging support (#2470)
Implement paging support from protofile.

This PR proposes creating a PagingConfig for a method in defined in the protofile iff

it follows the List Pagination guidelines and a GAPIC config for this method was not given.
Therefore, if a GAPIC config was given, and paging was not defined, then we wouldn't generate a PagingConfig because we should respect the absence of the GAPIC page_streaming value (absence of the page_streaming value is the only way to toggle OFF paging for a method.)

See baseline file for changes! The new additions to the changed no-config-generated baseline are existing LOC in the GAPIC-config-generated library baseline, so the diff between this new baseline and the existing java_library.baseline is even smaller.