[configgen] Handle singular resource entities (#2637)
[configgen] Handle singular resource entities

Generates a different entity name for singular resources than for their 
parent collections in gapic config generation. A singular resource is a 
different resource entirely from the parent collection.

For example, "users/*" is a collection resource for users, while 
"users/*/profile" is a singular resource for a user profile. The entity 
name of the collection resource would be `user`, and the entity name of 
the singular resource is `user_profile`.

Prefixes the singular resource entity name with its parent collection. 
This runs the risk of redundancy, but is clearer when the singular 
resource depends on its parent context.

Maintains the simple entity names for the path templates. These names 
only refer to collection resources since singular resources don't have a 
corresponding wildcard value.

Fixes #2633