Add GAPIC Bazel Extensions for Node.js
The extensions provide minimum level of functionality. It generates the output identical to what artman provides (i.e. has feature parity with artman).
It does not try to build and run the generated tests from withing bazel (like it does for Java and Go). This functionality may be added later.