Add GAPIC Bazel Extensions for Node.js (#2845)
The extensions provide minimum level of functionality. 

They generate the output identical to what artman provides (i.e. have feature parity with artman).

It does not try to build and run the generated tests from withing bazel (like it does for Java and Go). This functionality may be added later.