Make it possible to pass metadata to the operations client (#2856)
* Make it possible to pass metadata to the operations client

For googleads we inject `developer-token` and `login-customer-id` to all
Reques)ts. Previously to this patch, we had used a [monkey patch](https://git.io/fjrda)
to inject them in to the right place. However, ths patch was built
against an old version of both gax and the generated code and no longer
works for the newly generated google ads v2 protos.

At the core of the requirement is that both the main API client and,
where generated, the long running operations client, have injecatable
metadata so that we can provide them for LROs without patches.

This patch passes the metadata value passed in to the constructed client
along to the Long Running Operations client, which was not happening
before.

* baseline changes

* Force gax to 1.6.x

* Update bundle with gax in the Gemfile