Reconstruct PackagingConfig from other sources (gapic_yaml and protos) (#2862)
This makes `--package_yaml2` command line argument fully optional (for all languages). `--package_yaml2` is created by artman from `artman_yaml.` Not needing `--package_yaml2` as an argument effectively removes gapic-generator's "indirect dependency" on artman.

This is a prerequisite for Bazel generation for all languages which relay on `PackagingConfig` data (Ruby specifically, and all of the samples in all of the languages).