Switch to machine executor for Bazel builds (to fix tests flakiness) (#2909)
With a dedicated machine executor Bazel can determine properly the actual resources available (amount of CPUs and memory). 

When executed inside a Docker, Bazel sees the huge machine on which docker is running (i.e. stuff in `/protoc/cpuinfo` and `/proc/meminfo`) and assumes that all of those resources are available (while they are not, because Docker limits the access to them). 

As a result, when Bazel was executed from a Docker image, it tried to run tons of actions in parallel (assuming it had enough resources to do so), and circleci was simply killing it for consuming more resources than allowed. 

The previously employed `--local_resources` argument helped to mitigate the issue, but was not enough because it could control how much resources bazel and its direct tasks where allowed to consume, but the argument was not affecting subprocesses spawned by bazel's direct tasks (like javac Autovalue plugin, which was causing OutOfMemory exceptions and was the root cause of Bazel build tests flakiness).

With dedicated machine in place, all the tools used in build (Bazel itself and javac plugins) can see the real specs of the environment they are running on and "plan" accordingly (i.e. not fail with OutOfMemory and similar stuff).

Also upgrade bazel to the latest version (0.28.1).