Fix Resource Names annotations (#2979)
This is WIP

Add support for common resource names.
This does not fix all of the existing problems with the resource names.

Related fixes for `artman` and `protoc-java-resource-names-plugin` are pending.

**Config changes**
[library.proto](https://github.com/googleapis/gapic-generator/pull/2979/files#diff-a44629aaf921c03fdf47901df114cebf)  - relevant for gapic v1 and gapic v2 test cases
[library_gapic.yaml](https://github.com/googleapis/gapic-generator/pull/2979/files#diff-9f9ffb93516a846c5e39b7afc9e0c5de
) - relevant for gapic v1 test cases
[library_v2_gapic.yaml](https://github.com/googleapis/gapic-generator/pull/2979/files#diff-fc97e801b7222b7481527c795281e442) - relevant for gapic v2 test cases (basically just a fix of a bug)

The test cases try to do same in v1 and v2 worlds and check if the output produces same (or at least similar and binary compatible result).

The tests try to test:
- simple common resource name case (using `cloudresourcemanager.googleapis.com/Folder` as example):
  - referenced as `type` (`GetBookFromAnywhereRequest.folder`)
- common resource name containing the "calculated" parent in its pattern (using `locations.googleapis.com/Location` as an example) - notice **the difference** in generated code foe ach of the two cases below:
   - referenced as `chilt_type` (`GetBookFromArchiveRequest.parent`) - see `location_parent_oneof` in `library_gapic.yaml` for an analogous config in v1
   - referenced as `type` (`GetBookFromAnywhereRequest.place`) 
- language overrides (common_resource_name in C#) - (it was checked during real library generation, but no specific tests case added  here, will add soon)

**Output Changes**
The config changes produced output changes in the `<lang>_library.baseline` files for both of the sets of test cases (v1 and v2). Lang owners, please check that each of the `<lang>_library.baseline` file matches correctly the changes in the input (i.e. in library.proto and gapic yamls).